Basic imgui integration for Tools
with "showEditors" in the console you can get a simple window that
doesn't do much (but is intended to start the ingame editors you want,
like light editor, particle editor etc, that I plan to port from
doom3/mfc to d3bfg/imgui)

"Why another GUI toolkit?" because with imgui it's much easier/faster
to write GUIs than with cegui. The downside is that it's not as
flexible and themeable. So it's a good choice for ingame tools, but not
for the things a normal player will face, like the main menu.
Another advantage over cegui is that tools using imgui will even be
useful for projects that want to stick to the "old" flash GUI or
something else instead of cegui.