Fix crash when loading PNGs, fixes #59
The problem was that fileSystem->ReadFile() read the data to
io_s.buffer, so to release the data later
fileSystem->FreeFile(io_s.buffer); was called.
However, in between io_s.buffer has been changed by png_ReadData()
(p->buffer += length) and possibly by other png-internal functions
as well. Either way, when calling FreeFile() on it, it doesn't point to
the originally allocated address anymore, so  it causes a crash.

Fixed it by saving that address into an extra pointer and freeing that
later.