Restored much of the old ingame editor infrastructure, basic Light Editor
The Light Editor is a partly functional clone of the original one, but
incomplete. You can change the settings in real time (at least with the
color that seems to work as expected), but you can't save that to the
map yet. Might also be buggy.

Anyway, to try it, start a map, enter "g_editEntityMode 1" in the
console and "shoot" at a light.
The light editor window should turn up and you can play with the
settings.
Most of the settings are "Drags": You can click them and drag left/right
with the mouse to change the value.
Or double-click them and enter a value.

This uses ImGui instead of the old MFC crap.