CEGUI main menu can be closed with Esc when in game
For this idCEGUI::InjectSysEvent() is called explicitly from the
menu/console input handlers (ProcessEvent()/HandleGuiEvent())
and their return value depends on InjectSysEvents() return value.

Furthermore the CEGUI idMenuHandler_Shell::SetNextScreen() closes
the menu when it's called with "screen" set to SHELL_AREA_INVALID; this
seems to be D3BFG's way to tell the main menu to close.